+++ src/index.ts
@@ -1,27 +1,27 @@
 import { apply, isSupported } from './popover.js';
 
 interface PopoverToggleTargetElementInvoker {
-  popoverToggleTargetElement: HTMLElement | null;
-  popoverShowTargetElement: HTMLElement | null;
-  popoverHideTargetElement: HTMLElement | null;
+  popoverTargetElement: HTMLElement | null;
+  popoverTargetAction: 'toggle' | 'show' | 'hide';
 }
 
 declare global {
-  interface BeforeToggleEvent extends Event {
-    currentState: string;
+  interface ToggleEvent extends Event {
+    oldState: string;
     newState: string;
   }
   interface HTMLElement {
     popover: 'auto' | 'manual' | null;
     showPopover(): void;
     hidePopover(): void;
+    togglePopover(): void;
   }
   /* eslint-disable @typescript-eslint/no-empty-interface */
   interface HTMLButtonElement extends PopoverToggleTargetElementInvoker {}
   interface HTMLInputElement extends PopoverToggleTargetElementInvoker {}
   /* eslint-enable @typescript-eslint/no-empty-interface */
   interface Window {
-    BeforeToggleEvent: BeforeToggleEvent;
+    ToggleEvent: ToggleEvent;
   }
 }