+++ src/popover.ts
@@ -1,10 +1,11 @@
-import { openPopoverSelector, popoverInvokerSelector } from './data.js';
-import { observePopoversMutations } from './observer.js';
+import { ToggleEvent } from './events.js';
 import {
-  closestShadowPenetrating,
-  getInvokersFor,
-  hideOpenAutoPopovers,
-  setInvokerAriaExpanded,
+  hideAllPopoversUntil,
+  hidePopover,
+  lightDismissOpenPopovers,
+  popoverTargetAttributeActivationBehavior,
+  showPopover,
+  visibilityState,
 } from './popover-helpers.js';
 
 export function isSupported() {
@@ -15,81 +16,18 @@ export function isSupported() {
   );
 }
 
-function patchAttachShadow(callback: (shadowRoot: ShadowRoot) => void) {
-  const originalAttachShadow = Element.prototype.attachShadow;
-  Element.prototype.attachShadow = function (init) {
-    const shadowRoot = originalAttachShadow.call(this, init);
-    callback(shadowRoot);
-    return shadowRoot;
-  };
-}
-
 export function apply() {
-  const visibleElements = new WeakSet<HTMLElement>();
-  let lastFocusedElement: HTMLElement | null = null;
-
-  // https://whatpr.org/html/8221/popover.html#check-popover-validity
-  function checkPopoverValidity(
-    element: HTMLElement,
-    expectedToBeShowing: boolean,
-  ) {
-    if (element.popover !== 'auto' && element.popover !== 'manual') {
-      return false;
-    }
-    if (!element.isConnected) return false;
-    if (element instanceof HTMLDialogElement && element.hasAttribute('open')) {
-      return false;
-    }
-    if (expectedToBeShowing && !visibleElements.has(element)) return false;
-    if (!expectedToBeShowing && visibleElements.has(element)) return false;
-    if (document.fullscreenElement === element) return false;
-    return true;
-  }
-
-  function assertPopoverValidity(
-    element: HTMLElement,
-    expectedToBeShowing: boolean,
-  ) {
-    if (!checkPopoverValidity(element, expectedToBeShowing)) {
-      throw new DOMException(
-        'Cannot show or hide popover on invalid or already visible element',
-        'InvalidStateError',
-      );
-    }
-  }
-
-  interface BeforeToggleInit extends EventInit {
-    currentState: string;
-    newState: string;
-  }
-
-  class BeforeToggleEvent extends Event {
-    public currentState: string;
-    public newState: string;
-    constructor(
-      type: string,
-      {
-        currentState = '',
-        newState = '',
-        ...init
-      }: Partial<BeforeToggleInit> = {},
-    ) {
-      super(type, init);
-      this.currentState = String(currentState || '');
-      this.newState = String(newState || '');
-    }
-  }
-  window.BeforeToggleEvent = window.BeforeToggleEvent || BeforeToggleEvent;
+  window.ToggleEvent = window.ToggleEvent || ToggleEvent;
 
   Object.defineProperties(HTMLElement.prototype, {
     popover: {
       enumerable: true,
       configurable: true,
       get() {
+        if (!this.hasAttribute('popover')) return null;
         const value = (this.getAttribute('popover') || '').toLowerCase();
-        if (value === 'manual') return 'manual';
         if (value === '' || value == 'auto') return 'auto';
-        return null;
+        return 'manual';
       },
       set(value) {
         this.setAttribute('popover', value);
@@ -100,30 +38,7 @@ export function apply() {
       enumerable: true,
       configurable: true,
       value() {
-        // https://whatpr.org/html/8221/popover.html#show-popover
-        assertPopoverValidity(this, false);
-        const event = new BeforeToggleEvent('beforetoggle', {
-          cancelable: true,
-          currentState: 'closed',
-          newState: 'open',
-        });
-        if (!this.dispatchEvent(event)) return;
-        assertPopoverValidity(this, false);
-        this.classList.add(':open');
-        visibleElements.add(this);
-        const focusEl = this.hasAttribute('autofocus')
-          ? this
-          : this.querySelector('[autofocus]');
-        if (this.ownerDocument.activeElement instanceof HTMLElement) {
-          lastFocusedElement = this.ownerDocument.activeElement;
-        }
-        focusEl?.focus();
-        if (this.popover === 'auto') {
-          for (const invoker of getInvokersFor(this)) {
-            setInvokerAriaExpanded(invoker);
-          }
-          hideOpenAutoPopovers(this);
-        }
+        showPopover(this);
       },
     },
 
@@ -131,97 +46,105 @@ export function apply() {
       enumerable: true,
       configurable: true,
       value() {
-        // https://whatpr.org/html/8221/popover.html#hide-popover
-        assertPopoverValidity(this, true);
-        this.dispatchEvent(
-          new BeforeToggleEvent('beforetoggle', {
-            cancelable: false,
-            currentState: 'open',
-            newState: 'closed',
-          }),
-        );
-        assertPopoverValidity(this, true);
-        this.classList.remove(':open');
-        visibleElements.delete(this);
-        if (this.popover === 'auto') {
-          if (lastFocusedElement) {
-            lastFocusedElement.focus();
-            lastFocusedElement = null;
-          }
-          for (const invoker of getInvokersFor(this)) {
-            setInvokerAriaExpanded(invoker);
-          }
+        hidePopover(this, true, true);
+      },
+    },
+
+    togglePopover: {
+      enumerable: true,
+      configurable: true,
+      value(force: boolean) {
+        if (
+          (visibilityState.get(this) === 'showing' && force === undefined) ||
+          force === false
+        ) {
+          hidePopover(this, true, true);
+        } else if (force === undefined || force === true) {
+          showPopover(this);
         }
       },
     },
   });
 
-  const definePopoverTargetElementProperty = (name: string) => {
-    const invokersMap = new WeakMap<Element, Element>();
-    const invokerDescriptor: PropertyDescriptor &
-      ThisType<HTMLButtonElement | HTMLInputElement> = {
-      set(targetElement: unknown) {
-        if (targetElement === null) {
-          this.removeAttribute(name.toLowerCase());
-          invokersMap.delete(this);
-        } else if (!(targetElement instanceof Element)) {
-          throw new TypeError(`${name}Element must be an element or null`);
-        } else {
-          this.setAttribute(name.toLowerCase(), '');
-          invokersMap.set(this, targetElement);
-        }
+  const popoverTargetAssociatedElements = new WeakMap<Element, Element>();
+  function applyPopoverInvokerElementMixin(ElementClass: typeof HTMLElement) {
+    Object.defineProperties(ElementClass.prototype, {
+      popoverTargetElement: {
+        enumerable: true,
+        configurable: true,
+        set(targetElement: unknown) {
+          if (targetElement === null) {
+            this.removeAttribute('popovertarget');
+            popoverTargetAssociatedElements.delete(this);
+          } else if (!(targetElement instanceof Element)) {
+            throw new TypeError(
+              `popoverTargetElement must be an element or null`,
+            );
+          } else {
+            this.setAttribute('popovertarget', '');
+            popoverTargetAssociatedElements.set(this, targetElement);
+          }
+        },
+        get() {
+          if (this.localName !== 'button' && this.localName !== 'input') {
+            return null;
+          }
+          if (
+            this.localName === 'input' &&
+            this.type !== 'reset' &&
+            this.type !== 'image' &&
+            this.type !== 'button'
+          ) {
+            return null;
+          }
+          if (this.disabled) {
+            return null;
+          }
+          if (this.form && this.type === 'submit') {
+            return null;
+          }
+          const targetElement = popoverTargetAssociatedElements.get(this);
+          if (!targetElement?.isConnected) {
+            popoverTargetAssociatedElements.delete(this);
+          }
+          if (targetElement) {
+            return targetElement;
+          }
+          const root = this.getRootNode();
+          const idref = this.getAttribute('popovertarget');
+          if (
+            (root instanceof Document || root instanceof ShadowRoot) &&
+            idref
+          ) {
+            return root.getElementById(idref) || null;
+          }
+          return null;
+        },
       },
-      get() {
-        if (this.localName !== 'button' && this.localName !== 'input') {
-          return null;
-        }
-        if (
-          this.localName === 'input' &&
-          this.type !== 'reset' &&
-          this.type !== 'image' &&
-          this.type !== 'button'
-        ) {
-          return null;
-        }
-        if (this.disabled) {
-          return null;
-        }
-        if (this.form && this.type === 'submit') {
-          return null;
-        }
-        const targetElement = invokersMap.get(this);
-        if (!targetElement?.isConnected) {
-          invokersMap.delete(this);
-        }
-        if (targetElement) {
-          return targetElement;
-        }
-        const root = this.getRootNode();
-        const idref = this.getAttribute(name.toLowerCase());
-        if ((root instanceof Document || root instanceof ShadowRoot) && idref) {
-          return root.getElementById(idref) || null;
-        }
-        return null;
+      popoverTargetAction: {
+        enumerable: true,
+        configurable: true,
+        get() {
+          const value = (
+            this.getAttribute('popovertargetaction') || ''
+          ).toLowerCase();
+          if (value === 'show' || value === 'hide') return value;
+          return 'toggle';
+        },
+        set(value) {
+          this.setAttribute('popovertargetaction', value);
+        },
       },
-    };
-    Object.defineProperty(
-      HTMLButtonElement.prototype,
-      `${name}Element`,
-      invokerDescriptor,
-    );
-    Object.defineProperty(
-      HTMLInputElement.prototype,
-      `${name}Element`,
-      invokerDescriptor,
-    );
-  };
+    });
+  }
 
-  definePopoverTargetElementProperty('popoverToggleTarget');
-  definePopoverTargetElementProperty('popoverShowTarget');
-  definePopoverTargetElementProperty('popoverHideTarget');
+  applyPopoverInvokerElementMixin(HTMLButtonElement);
+  applyPopoverInvokerElementMixin(HTMLInputElement);
 
-  const onClick = (event: Event) => {
-    const target = event.target;
+  const handleInvokerActivation = (event: Event) => {
+    if (!event.isTrusted) return;
+    // Composed path allows us to find the target within shadowroots
+    const target = event.composedPath()[0] as HTMLElement;
     if (!(target instanceof Element) || target?.shadowRoot) {
       return;
     }
@@ -229,57 +152,27 @@ export function apply() {
     if (!(root instanceof ShadowRoot || root instanceof Document)) {
       return;
     }
-    const invoker = target.closest(popoverInvokerSelector);
-    let popoverTargetElement: HTMLElement | null = null;
-    if (
-      invoker instanceof HTMLButtonElement ||
-      invoker instanceof HTMLInputElement
-    ) {
-      if (invoker.popoverToggleTargetElement) {
-        popoverTargetElement = invoker.popoverToggleTargetElement;
-        if (popoverTargetElement) {
-          if (visibleElements.has(popoverTargetElement)) {
-            popoverTargetElement.hidePopover();
-          } else {
-            popoverTargetElement.showPopover();
-          }
-        }
-      } else if (invoker.popoverShowTargetElement) {
-        popoverTargetElement = invoker.popoverShowTargetElement;
-        if (
-          popoverTargetElement &&
-          !visibleElements.has(popoverTargetElement)
-        ) {
-          popoverTargetElement.showPopover();
-        }
-      } else if (invoker.popoverHideTargetElement) {
-        popoverTargetElement = invoker.popoverHideTargetElement;
-        if (popoverTargetElement && visibleElements.has(popoverTargetElement)) {
-          popoverTargetElement.hidePopover();
-        }
-      }
+    const invoker = target.closest('[popovertargetaction],[popovertarget]');
+    if (invoker) {
+      popoverTargetAttributeActivationBehavior(invoker as HTMLButtonElement);
+      return;
     }
-    hideOpenAutoPopovers(
-      popoverTargetElement ||
-        closestShadowPenetrating(openPopoverSelector, target),
-    );
   };
 
   const onKeydown = (event: Event) => {
     const key = (event as KeyboardEvent).key;
-    if (key === 'Escape' || key === 'Esc') {
-      hideOpenAutoPopovers();
+    const target = event.target as Element;
+    if (target && (key === 'Escape' || key === 'Esc')) {
+      hideAllPopoversUntil(target.ownerDocument, true, true);
     }
   };
 
-  const addOnClickEventListener = (root: Document | ShadowRoot) => {
-    root.addEventListener('click', onClick);
+  const addEventListeners = (root: Document | ShadowRoot) => {
+    root.addEventListener('click', handleInvokerActivation);
     root.addEventListener('keydown', onKeydown);
+    root.addEventListener('pointerdown', lightDismissOpenPopovers);
+    root.addEventListener('pointerup', lightDismissOpenPopovers);
   };
 
-  observePopoversMutations(document);
-  addOnClickEventListener(document);
-
-  patchAttachShadow(observePopoversMutations);
-  patchAttachShadow(addOnClickEventListener);
+  addEventListeners(document);
 }